Move responses-compiler-only code to responses compiler
