+++ source/libocto/compilers/openai-shared.ts
@@ -1,0 +1,9 @@
+import type OpenAI from "openai";
+import type { CompilerModalities } from "./compiler-interface.ts";
+
+export type OpenAICompilerModel = {
+  client: OpenAI;
+  model: string;
+  modalities?: CompilerModalities;
+  reasoningEffort?: "low" | "medium" | "high";
+};
